Refactor Google Maps API integration
- Conditionally export web or stub implementation for better platform compatibility.
- Removed deprecated code related to direct API initialization.
