+++ pasada_admin_application/lib/maps/google_maps_api_stub.dart
@@ -1,0 +1,7 @@
+/// No-op implementation for non-web platforms.
+class GoogleMapsApiInitializer {
+  /// Does nothing on non-web platforms.
+  static void initialize() {
+    // No-op
+  }
+}
